images: Add tests about osm images
